Fixed general error handling for whole program
In chat_interface, the whole thing is in a try-except block, so the
except should in theory capture anything wrong with the whole chatbot.
However, we kept running into silent failures, and this wasted a lot
of time for debugging. I realized the core of the issue is that when
the model is set to streaming=True, it expects a generator to be returned,
however, currently we return an error string. So now added a check for
streaming=True, and if so, we yield that error string. Fixed!